moved some hardcoded html into the pages, setup typescript modules
