+++ src/models/stanfordbunny.ts
@@ -1,7 +1,6 @@
 
 
-
-class StanfordBunny {
+export class StanfordBunny {
 
     public bunnyIndices: number[] = [
         0, 1, 2, 3, 4, 5, 6, 7, 8, 9, 10, 11, 12, 13, 14, 15, 16, 17, 18, 17, 19, 20, 21, 22, 23, 24, 25, 26, 27, 28, 23, 29, 30, 30, 29, 31, 31,
@@ -11715,762 +11714,4 @@ class StanfordBunny {
         0.007019, -0.304784, -0.618703, 0.037066, 0.078818, -0.996341, -0.033053
     ];
 }
-
-
-
-class SphereModel {
-
-    public sphereIndices: number[] = [
-        0, 1, 2, 3, 4, 5, 0, 2, 6, 0, 6, 7, 0, 7, 8, 3, 5, 9, 10, 11, 12, 13, 14, 15, 16, 17, 18, 19, 20, 21, 3, 9, 22, 10, 12, 23, 13, 15, 24, 16, 18,
-        25, 19, 21, 26, 27, 28, 29, 30, 31, 32, 33, 34, 35, 36, 37, 38, 39, 40, 41, 42, 11, 10, 43, 44, 42, 2, 45, 43, 42, 44, 11, 44, 46, 11,
-        43, 45, 44, 45, 47, 44, 44, 47, 46, 47, 48, 46, 2, 1, 45, 1, 49, 45, 45, 49, 47, 49, 4, 47, 47, 4, 48, 4, 3, 48, 50, 51, 19, 52, 53, 50,
-        5, 54, 52, 50, 53, 51, 53, 55, 51, 52, 54, 53, 54, 56, 53, 53, 56, 55, 56, 8, 55, 5, 4, 54, 4, 49, 54, 54, 49, 56, 49, 1, 56, 56, 1, 8, 1,
-        0, 8, 57, 14, 13, 58, 59, 57, 6, 60, 58, 57, 59, 14, 59, 61, 14, 58, 60, 59, 60, 62, 59, 59, 62, 61, 62, 63, 61, 6, 2, 60, 2, 43, 60, 60,
-        43, 62, 43, 42, 62, 62, 42, 63, 42, 10, 63, 64, 17, 16, 65, 66, 64, 7, 67, 65, 64, 66, 17, 66, 68, 17, 65, 67, 66, 67, 69, 66, 66, 69,
-        68, 69, 70, 68, 7, 6, 67, 6, 58, 67, 67, 58, 69, 58, 57, 69, 69, 57, 70, 57, 13, 70, 51, 20, 19, 55, 71, 51, 8, 72, 55, 51, 71, 20, 71,
-        73, 20, 55, 72, 71, 72, 74, 71, 71, 74, 73, 74, 75, 73, 8, 7, 72, 7, 65, 72, 72, 65, 74, 65, 64, 74, 74, 64, 75, 64, 16, 75, 76, 77, 39,
-        78, 79, 76, 9, 80, 78, 76, 79, 77, 79, 81, 77, 78, 80, 79, 80, 82, 79, 79, 82, 81, 82, 26, 81, 9, 5, 80, 5, 52, 80, 80, 52, 82, 52, 50,
-        82, 82, 50, 26, 50, 19, 26, 83, 84, 27, 85, 86, 83, 12, 87, 85, 83, 86, 84, 86, 88, 84, 85, 87, 86, 87, 89, 86, 86, 89, 88, 89, 22, 88,
-        12, 11, 87, 11, 46, 87, 87, 46, 89, 46, 48, 89, 89, 48, 22, 48, 3, 22, 90, 91, 30, 92, 93, 90, 15, 94, 92, 90, 93, 91, 93, 95, 91, 92,
-        94, 93, 94, 96, 93, 93, 96, 95, 96, 23, 95, 15, 14, 94, 14, 61, 94, 94, 61, 96, 61, 63, 96, 96, 63, 23, 63, 10, 23, 97, 98, 33, 99, 100,
-        97, 18, 101, 99, 97, 100, 98, 100, 102, 98, 99, 101, 100, 101, 103, 100, 100, 103, 102, 103, 24, 102, 18, 17, 101, 17, 68, 101, 101,
-        68, 103, 68, 70, 103, 103, 70, 24, 70, 13, 24, 104, 105, 36, 106, 107, 104, 21, 108, 106, 104, 107, 105, 107, 109, 105, 106, 108,
-        107, 108, 110, 107, 107, 110, 109, 110, 25, 109, 21, 20, 108, 20, 73, 108, 108, 73, 110, 73, 75, 110, 110, 75, 25, 75, 16, 25, 84,
-        28, 27, 88, 111, 84, 22, 112, 88, 84, 111, 28, 111, 113, 28, 88, 112, 111, 112, 114, 111, 111, 114, 113, 114, 115, 113, 22, 9, 112,
-        9, 78, 112, 112, 78, 114, 78, 76, 114, 114, 76, 115, 76, 39, 115, 91, 31, 30, 95, 116, 91, 23, 117, 95, 91, 116, 31, 116, 118, 31,
-        95, 117, 116, 117, 119, 116, 116, 119, 118, 119, 120, 118, 23, 12, 117, 12, 85, 117, 117, 85, 119, 85, 83, 119, 119, 83, 120, 83,
-        27, 120, 98, 34, 33, 102, 121, 98, 24, 122, 102, 98, 121, 34, 121, 123, 34, 102, 122, 121, 122, 124, 121, 121, 124, 123, 124, 125,
-        123, 24, 15, 122, 15, 92, 122, 122, 92, 124, 92, 90, 124, 124, 90, 125, 90, 30, 125, 105, 37, 36, 109, 126, 105, 25, 127, 109, 105,
-        126, 37, 126, 128, 37, 109, 127, 126, 127, 129, 126, 126, 129, 128, 129, 130, 128, 25, 18, 127, 18, 99, 127, 127, 99, 129, 99, 97,
-        129, 129, 97, 130, 97, 33, 130, 77, 40, 39, 81, 131, 77, 26, 132, 81, 77, 131, 40, 131, 133, 40, 81, 132, 131, 132, 134, 131, 131,
-        134, 133, 134, 135, 133, 26, 21, 132, 21, 106, 132, 132, 106, 134, 106, 104, 134, 134, 104, 135, 104, 36, 135, 136, 137, 138, 139,
-        140, 136, 29, 141, 139, 136, 140, 137, 140, 142, 137, 139, 141, 140, 141, 143, 140, 140, 143, 142, 143, 41, 142, 29, 28, 141, 28,
-        113, 141, 141, 113, 143, 113, 115, 143, 143, 115, 41, 115, 39, 41, 144, 136, 138, 145, 146, 144, 32, 147, 145, 144, 146, 136, 146,
-        139, 136, 145, 147, 146, 147, 148, 146, 146, 148, 139, 148, 29, 139, 32, 31, 147, 31, 118, 147, 147, 118, 148, 118, 120, 148, 148,
-        120, 29, 120, 27, 29, 149, 144, 138, 150, 151, 149, 35, 152, 150, 149, 151, 144, 151, 145, 144, 150, 152, 151, 152, 153, 151, 151,
-        153, 145, 153, 32, 145, 35, 34, 152, 34, 123, 152, 152, 123, 153, 123, 125, 153, 153, 125, 32, 125, 30, 32, 154, 149, 138, 155,
-        156, 154, 38, 157, 155, 154, 156, 149, 156, 150, 149, 155, 157, 156, 157, 158, 156, 156, 158, 150, 158, 35, 150, 38, 37, 157, 37,
-        128, 157, 157, 128, 158, 128, 130, 158, 158, 130, 35, 130, 33, 35, 137, 154, 138, 142, 159, 137, 41, 160, 142, 137, 159, 154, 159,
-        155, 154, 142, 160, 159, 160, 161, 159, 159, 161, 155, 161, 38, 155, 41, 40, 160, 40, 133, 160, 160, 133, 161, 133, 135, 161, 161,
-        135, 38, 135, 36, 38
-    ];
-
-    public sphereVertices: number[] = [
-        0, -1, 0, 0, -1, 0, 0.203181, -0.96795, 0.147618, 0.210944, -0.965392, 0.153264, -0.077607, -0.96795, 0.238853, -0.080569,
-        -0.965392, 0.247963, 0.723607, -0.44722, 0.525725, 0.723594, -0.447218, 0.525712, 0.609547, -0.657519, 0.442856,
-        0.604205, -0.664968, 0.438978, 0.812729, -0.502301, 0.295238, 0.81518, -0.5038, 0.285714, -0.251147, -0.967949,
-        0, -0.26075, -0.965392, 0, -0.077607, -0.96795, -0.238853, -0.080569, -0.965392, -0.247963, 0.203181, -0.96795, -0.147618,
-        0.210944, -0.965392, -0.153264, 0.860698, -0.251151, 0.442858, 0.864986, -0.243049, 0.438978, -0.276388, -0.44722,
-        0.850649, -0.276376, -0.447218, 0.850642, -0.029639, -0.502302, 0.864184, -0.019837, -0.5038, 0.863552, -0.155215,
-        -0.251152, 0.955422, -0.150212, -0.243049, 0.958281, -0.894426, -0.447216, 0, -0.894406, -0.447188, 0, -0.831051,
-        -0.502299, 0.238853, -0.827448, -0.5038, 0.247963, -0.956626, -0.251149, 0.147618, -0.957823, -0.243049, 0.153264,
-        -0.276388, -0.44722, -0.850649, -0.276376, -0.447218, -0.850642, -0.483971, -0.502302, -0.716565, -0.491531, -0.5038,
-        -0.710288, -0.436007, -0.251152, -0.864188, -0.441725, -0.243049, -0.863582, 0.723607, -0.44722, -0.525725, 0.723594,
-        -0.447218, -0.525712, 0.531941, -0.502302, -0.681712, 0.523637, -0.5038, -0.686972, 0.687159, -0.251152, -0.681715,
-        0.684805, -0.243049, -0.686972, 0.687159, -0.251152, 0.681715, 0.684805, -0.243049, 0.686972, -0.436007, -0.251152,
-        0.864188, -0.441725, -0.243049, 0.863582, -0.956626, -0.251149, -0.147618, -0.957823, -0.243049, -0.153264, -0.155215,
-        -0.251152, -0.955422, -0.150212, -0.243049, -0.958281, 0.860698, -0.251151, -0.442858, 0.864986, -0.243049, -0.438978,
-        0.276388, 0.44722, 0.850649, 0.276376, 0.447218, 0.850642, 0.483971, 0.502302, 0.716565, 0.491531, 0.5038, 0.710288,
-        0.232822, 0.657519, 0.716563, 0.230781, 0.664968, 0.710288, -0.723607, 0.44722, 0.525725, -0.723594, 0.447218, 0.525712,
-        -0.531941, 0.502302, 0.681712, -0.523637, 0.5038, 0.686972, -0.609547, 0.657519, 0.442856, -0.604205, 0.664968,
-        0.438978, -0.723607, 0.44722, -0.525725, -0.723594, 0.447218, -0.525712, -0.812729, 0.502301, -0.295238, -0.81518,
-        0.5038, -0.285714, -0.609547, 0.657519, -0.442856, -0.604205, 0.664968, -0.438978, 0.276388, 0.44722, -0.850649,
-        0.276376, 0.447218, -0.850642, 0.029639, 0.502302, -0.864184, 0.019837, 0.5038, -0.863552, 0.232822, 0.657519, -0.716563,
-        0.230781, 0.664968, -0.710288, 0.894426, 0.447216, 0, 0.894406, 0.447188, 0, 0.831051, 0.502299, -0.238853, 0.827448,
-        0.5038, -0.247963, 0.753442, 0.657515, 0, 0.746849, 0.664968, 0, -0.232822, -0.657519, 0.716563, -0.230781, -0.664968,
-        0.710288, -0.162456, -0.850654, 0.499995, -0.16245, -0.850642, 0.499985, 0.05279, -0.723612, 0.688185, 0.059206,
-        -0.727531, 0.683493, 0.138199, -0.894429, 0.425321, 0.140629, -0.890408, 0.432844, 0.262869, -0.525738, 0.809012,
-        0.262856, -0.525712, 0.808985, 0.361805, -0.723611, 0.587779, 0.353832, -0.727531, 0.587756, 0.531941, -0.502302,
-        0.681712, 0.523637, -0.5038, 0.686972, 0.425323, -0.850654, 0.309011, 0.425306, -0.850642, 0.309, 0.812729, -0.502301,
-        -0.295238, 0.81518, -0.5038, -0.285714, 0.609547, -0.657519, -0.442856, 0.604205, -0.664968, -0.438978, 0.850648,
-        -0.525736, 0, 0.850642, -0.525712, 0, 0.670817, -0.723611, -0.162457, 0.668325, -0.727531, -0.154881, 0.670818, -0.72361,
-        0.162458, 0.668325, -0.727531, 0.154881, 0.425323, -0.850654, -0.309011, 0.425306, -0.850642, -0.309, 0.447211,
-        -0.894428, 1E-06, 0.455123, -0.890408, 0, -0.753442, -0.657515, 0, -0.746849, -0.664968, 0, -0.52573, -0.850652, 0,
-        -0.525712, -0.850642, 0, -0.638195, -0.723609, 0.262864, -0.631733, -0.727531, 0.267495, -0.361801, -0.894428, 0.262864,
-        -0.368206, -0.890408, 0.267495, -0.688189, -0.525736, 0.499997, -0.688162, -0.525712, 0.499985, -0.447211, -0.72361,
-        0.525729, -0.449629, -0.727531, 0.518143, -0.483971, -0.502302, 0.716565, -0.491531, -0.5038, 0.710288, -0.232822,
-        -0.657519, -0.716563, -0.230781, -0.664968, -0.710288, -0.162456, -0.850654, -0.499995, -0.16245, -0.850642, -0.499985,
-        -0.447211, -0.723612, -0.525727, -0.449629, -0.727531, -0.518143, -0.361801, -0.894429, -0.262863, -0.368206, -0.890408,
-        -0.267495, -0.688189, -0.525736, -0.499997, -0.688162, -0.525712, -0.499985, -0.638195, -0.723609, -0.262863, -0.631733,
-        -0.727531, -0.267495, -0.831051, -0.502299, -0.238853, -0.827448, -0.5038, -0.247963, 0.361803, -0.723612, -0.587779,
-        0.353832, -0.727531, -0.587756, 0.138197, -0.894429, -0.425321, 0.140629, -0.890408, -0.432844, 0.262869, -0.525738,
-        -0.809012, 0.262856, -0.525712, -0.808985, 0.052789, -0.723611, -0.688186, 0.059206, -0.727531, -0.683493, -0.029639,
-        -0.502302, -0.864184, -0.019837, -0.5038, -0.863552, 0.956626, 0.251149, 0.147618, 0.957823, 0.243049, 0.153264,
-        0.956626, 0.251149, -0.147618, 0.957823, 0.243049, -0.153264, 0.951058, 0, 0.309013, 0.951048, 0, 0.309, 1, 0, 0, 0.999939,
-        -0.00885, 0, 0.947213, -0.276396, 0.162458, 0.949614, -0.272408, 0.154881, 0.951058, 0, -0.309013, 0.951048, 0, -0.309,
-        0.947213, -0.276396, -0.162458, 0.949614, -0.272408, -0.154881, 0.155215, 0.251152, 0.955422, 0.150212, 0.243049,
-        0.958281, 0.436007, 0.251152, 0.864188, 0.441725, 0.243049, 0.863582, 0, 0, 1, 0, 0, 1, 0.309017, 0, 0.951056, 0.309,
-        -0.00885, 0.951018, 0.138199, -0.276398, 0.951055, 0.146123, -0.272408, 0.950987, 0.587786, 0, 0.809017, 0.587756,
-        0, 0.809015, 0.447216, -0.276398, 0.850648, 0.440748, -0.272408, 0.855251, -0.860698, 0.251151, 0.442858, -0.864986,
-        0.243049, 0.438978, -0.687159, 0.251152, 0.681715, -0.684805, 0.243049, 0.686972, -0.951058, 0, 0.309013, -0.951048,
-        0, 0.309, -0.809018, 0, 0.587783, -0.808985, -0.00885, 0.587756, -0.861803, -0.276396, 0.425324, -0.85931, -0.272408,
-        0.432844, -0.587786, 0, 0.809017, -0.587756, 0, 0.809015, -0.670819, -0.276397, 0.688191, -0.677206, -0.272408, 0.683493,
-        -0.687159, 0.251152, -0.681715, -0.684805, 0.243049, -0.686972, -0.860698, 0.251151, -0.442858, -0.864986, 0.243049,
-        -0.438978, -0.587786, 0, -0.809017, -0.587756, 0, -0.809015, -0.809018, 0, -0.587783, -0.808985, -0.00885, -0.587756,
-        -0.670819, -0.276397, -0.688191, -0.677206, -0.272408, -0.683493, -0.951058, 0, -0.309013, -0.951048, 0, -0.309,
-        -0.861803, -0.276396, -0.425324, -0.85931, -0.272408, -0.432844, 0.436007, 0.251152, -0.864188, 0.441725, 0.243049,
-        -0.863582, 0.155215, 0.251152, -0.955422, 0.150212, 0.243049, -0.958281, 0.587786, 0, -0.809017, 0.587756, 0, -0.809015,
-        0.309017, 0, -0.951056, 0.309, -0.00885, -0.951018, 0.447216, -0.276398, -0.850648, 0.440748, -0.272408, -0.855251,
-        0, 0, -1, 0, 0, -1, 0.138199, -0.276398, -0.951055, 0.146123, -0.272408, -0.950987, 0.67082, 0.276396, 0.68819, 0.677206,
-        0.272408, 0.683493, 0.809019, -2E-06, 0.587783, 0.808985, 0.00885, 0.587756, 0.688189, 0.525736, 0.499997, 0.688162,
-        0.525712, 0.499985, 0.861804, 0.276394, 0.425323, 0.85931, 0.272408, 0.432844, 0.831051, 0.502299, 0.238853, 0.827448,
-        0.5038, 0.247963, -0.447216, 0.276397, 0.850648, -0.440748, 0.272408, 0.855251, -0.309017, -1E-06, 0.951056, -0.309,
-        0.00885, 0.951018, -0.262869, 0.525738, 0.809012, -0.262856, 0.525712, 0.808985, -0.138199, 0.276397, 0.951055,
-        -0.146123, 0.272408, 0.950987, 0.029639, 0.502302, 0.864184, 0.019837, 0.5038, 0.863552, -0.947213, 0.276396, -0.162458,
-        -0.949614, 0.272408, -0.154881, -1, 1E-06, 0, -0.999939, 0.00885, 0, -0.850648, 0.525736, 0, -0.850642, 0.525712, 0,
-        -0.947213, 0.276397, 0.162458, -0.949614, 0.272408, 0.154881, -0.812729, 0.502301, 0.295238, -0.81518, 0.5038, 0.285714,
-        -0.138199, 0.276397, -0.951055, -0.146123, 0.272408, -0.950987, -0.309016, 0, -0.951057, -0.309, 0.00885, -0.951018,
-        -0.262869, 0.525738, -0.809012, -0.262856, 0.525712, -0.808985, -0.447215, 0.276397, -0.850649, -0.440748, 0.272408,
-        -0.855251, -0.531941, 0.502302, -0.681712, -0.523637, 0.5038, -0.686972, 0.861804, 0.276396, -0.425322, 0.85931,
-        0.272408, -0.432844, 0.809019, 0, -0.587782, 0.808985, 0.00885, -0.587756, 0.688189, 0.525736, -0.499997, 0.688162,
-        0.525712, -0.499985, 0.670821, 0.276397, -0.688189, 0.677206, 0.272408, -0.683493, 0.483971, 0.502302, -0.716565,
-        0.491531, 0.5038, -0.710288, 0.077607, 0.96795, 0.238853, 0.080569, 0.965392, 0.247963, 0.251147, 0.967949, 0, 0.26075,
-        0.965392, 0, 0, 1, 0, 0, 1, 0, 0.162456, 0.850654, 0.499995, 0.16245, 0.850642, 0.499985, 0.3618, 0.894429, 0.262863,
-        0.368206, 0.890408, 0.267495, 0.447209, 0.723612, 0.525728, 0.449629, 0.727531, 0.518143, 0.52573, 0.850652, 0, 0.525712,
-        0.850642, 0, 0.638194, 0.72361, 0.262864, 0.631733, 0.727531, 0.267495, -0.203181, 0.96795, 0.147618, -0.210944,
-        0.965392, 0.153264, -0.425323, 0.850654, 0.309011, -0.425306, 0.850642, 0.309, -0.138197, 0.89443, 0.425319, -0.140629,
-        0.890408, 0.432844, -0.361804, 0.723612, 0.587778, -0.353832, 0.727531, 0.587725, -0.05279, 0.723612, 0.688185,
-        -0.059206, 0.727531, 0.683493, -0.203181, 0.96795, -0.147618, -0.210944, 0.965392, -0.153264, -0.425323, 0.850654,
-        -0.309011, -0.425306, 0.850642, -0.309, -0.44721, 0.894429, 0, -0.455123, 0.890408, 0, -0.670817, 0.723611, -0.162457,
-        -0.668325, 0.727531, -0.154881, -0.670817, 0.723611, 0.162457, -0.668325, 0.727531, 0.154881, 0.077607, 0.96795,
-        -0.238853, 0.080569, 0.965392, -0.247963, 0.162456, 0.850654, -0.499995, 0.16245, 0.850642, -0.499985, -0.138197,
-        0.89443, -0.425319, -0.140629, 0.890408, -0.432844, -0.05279, 0.723612, -0.688185, -0.059206, 0.727531, -0.683493,
-        -0.361804, 0.723612, -0.587778, -0.353832, 0.727531, -0.587725, 0.3618, 0.894429, -0.262863, 0.368206, 0.890408,
-        -0.267495, 0.638194, 0.72361, -0.262864, 0.631733, 0.727531, -0.267495, 0.447209, 0.723612, -0.525728, 0.449629,
-        0.727531, -0.518143,
-    ];
-}
-
-
-class TriangleModel {
-
-    public indices: number[] = [0, 1, 2];
-    public vertices: number[] = [0.0, 0.75, 0.0, 1.0, 0.0, 0.0, -0.75, -0.75, 0.0, 0.0, 1.0, 0.0, 0.75, -0.75, 0.0, 0.0, 0.0, 1.0];
-}
-
-
-class WebGlHost {
-
-    public gl: WebGLRenderingContext | undefined | null;
-
-    public vertices: number[];
-    public indices: number[];
-
-    public vertexShaderCode: string;
-    public fragmentShaderCode: string;
-    public cameraPosition: number[] = [0.0, 0.0, 0.0];
-    public rotation = 0;
-    public shaderProgram: WebGLProgram | undefined | null;
-
-    public pageString: string;
-
-    constructor(vertices: number[], indices: number[], vertexShaderCode: string, fragmentShaderCode: string, pageString: string) {
-
-        this.vertices = vertices;
-        this.indices = indices;
-
-        this.vertexShaderCode = vertexShaderCode;
-        this.fragmentShaderCode = fragmentShaderCode;
-        this.pageString = pageString;
-        this.initialiseWebGL();
-        this.addButtonListener();
-    }
-
-    private initialiseWebGL(): void {
-
-        const canvas: HTMLElement | null = document.getElementById("webGLCanvas");
-
-        if (canvas && canvas instanceof HTMLCanvasElement) {
-
-            this.gl = canvas.getContext("webgl");
-
-            if (!this.gl) {
-                return;
-            }
-
-            var gl: WebGLRenderingContext = this.gl;
-
-            gl.clearColor(0.2, 0.2, 0.2, 1.0);
-            gl.enable(gl.DEPTH_TEST);
-            gl.clear(gl.COLOR_BUFFER_BIT);
-            gl.viewport(0, 0, canvas.width, canvas.height);
-            this.loadBuffers();
-            this.loadShaders();
-        }
-    }
-
-    public addButtonListener(): void {
-
-        var button = document.getElementById("update-button")
-
-        button?.addEventListener("click", () => {
-            this.clickEvent();
-        })
-    }
-
-    public clickEvent(): void {
-
-        switch (this.pageString) {
-            case "loading":
-                this.reloadBuffers();
-                return;
-
-            case "camera":
-                this.reloadVertexShader();
-                return;
-
-            case "lighting":
-                this.reloadPixelShader();
-                return;
-        }
-    }
-
-    public renderCycle(): void {
-
-        if (!this.gl) {
-            return;
-        }
-
-        var gl: WebGLRenderingContext = this.gl;
-
-        gl.clear(gl.COLOR_BUFFER_BIT | gl.DEPTH_BUFFER_BIT);
-        gl.drawElements(gl.TRIANGLES, this.indices.length, gl.UNSIGNED_SHORT, 0);
-    }
-
-    toNumArray(strings: string[]): number[] {
-
-        var ans: number[] = new Array<number>();
-
-        strings.forEach(function (s) {
-            ans.push(+s);
-        })
-
-        return ans;
-    }
-
-    loadBuffers(): void {
-
-        if (!this.gl) {
-            return;
-        }
-
-        var gl: WebGLRenderingContext = this.gl;
-
-        var vertex_buffer = gl.createBuffer();
-        gl.bindBuffer(gl.ARRAY_BUFFER, vertex_buffer);
-        gl.bufferData(gl.ARRAY_BUFFER, new Float32Array(this.vertices), gl.STATIC_DRAW);
-        gl.bindBuffer(gl.ARRAY_BUFFER, null);
-
-        var Index_Buffer = gl.createBuffer();
-        gl.bindBuffer(gl.ELEMENT_ARRAY_BUFFER, Index_Buffer);
-        gl.bufferData(gl.ELEMENT_ARRAY_BUFFER, new Uint16Array(this.indices), gl.STATIC_DRAW);
-        gl.bindBuffer(gl.ELEMENT_ARRAY_BUFFER, null);
-
-        /*======= Associating shaders to buffer objects =======*/
-
-        gl.bindBuffer(gl.ARRAY_BUFFER, vertex_buffer);
-        gl.bindBuffer(gl.ELEMENT_ARRAY_BUFFER, Index_Buffer);
-    }
-
-    loadShaders(): void {
-
-        if (!this.gl) {
-            return;
-        }
-
-        var gl: WebGLRenderingContext = this.gl;
-
-        var vertShader: WebGLShader | null = gl.createShader(gl.VERTEX_SHADER);
-
-        if (!vertShader) {
-            return;
-        }
-
-        gl.shaderSource(vertShader, this.vertexShaderCode);
-        gl.compileShader(vertShader);
-
-        var fragShader: WebGLShader | null = gl.createShader(gl.FRAGMENT_SHADER);
-
-        if (!fragShader) {
-            return;
-        }
-
-        gl.shaderSource(fragShader, this.fragmentShaderCode);
-        gl.compileShader(fragShader);
-
-
-        this.shaderProgram = gl.createProgram();
-
-        if (!this.shaderProgram) {
-            return;
-        }
-
-        gl.attachShader(this.shaderProgram, vertShader);
-        gl.attachShader(this.shaderProgram, fragShader);
-        gl.linkProgram(this.shaderProgram);
-        gl.useProgram(this.shaderProgram);
-    }
-
-    bind_a_position(): void {
-
-        if (!this.gl) {
-            return;
-        }
-
-        var gl: WebGLRenderingContext = this.gl;
-
-        if (!this.shaderProgram) {
-            return;
-        }
-
-        var coord = gl.getAttribLocation(this.shaderProgram, "a_position");
-        gl.vertexAttribPointer(coord, 3, gl.FLOAT, false, 6 * 4, 0);
-        gl.enableVertexAttribArray(coord);
-        this.renderCycle();
-    }
-
-    bind_a_normal(): void {
-
-        if (!this.gl) {
-            return;
-        }
-
-        var gl: WebGLRenderingContext = this.gl;
-
-        if (!this.shaderProgram) {
-            return;
-        }
-
-        var coord = gl.getAttribLocation(this.shaderProgram, "a_normal");
-        gl.vertexAttribPointer(coord, 3, gl.FLOAT, false, 6 * 4, 3 * 4);
-        gl.enableVertexAttribArray(coord);
-        this.renderCycle();
-    }
-
-
-    bind_a_colour() {
-
-        if (!this.gl) {
-            return;
-        }
-
-        var gl: WebGLRenderingContext = this.gl;
-
-        if (!this.shaderProgram) {
-            return;
-        }
-
-        var coord = gl.getAttribLocation(this.shaderProgram, "a_colour");
-        gl.vertexAttribPointer(coord, 3, gl.FLOAT, false, 6 * 4, 3 * 4);
-        gl.enableVertexAttribArray(coord);
-    }
-
-    updateSimpleCameraPosition() {
-
-        if (!this.gl) {
-            return;
-        }
-
-        var gl: WebGLRenderingContext = this.gl;
-
-        if (!this.shaderProgram) {
-            return;
-        }
-
-        var uCamPosLocation = gl.getUniformLocation(this.shaderProgram, "u_cameraPosition");
-        gl.uniform3f(uCamPosLocation,
-            this.cameraPosition[0],
-            this.cameraPosition[1],
-            this.cameraPosition[2]);
-    }
-
-    public updateCameraPositionOnKeyUp(event: KeyboardEvent) {
-
-        if (this.moveCamera(event.key)) {
-            this.updateSimpleCameraPosition();
-        }
-        this.renderCycle();
-    }
-
-    public moveCamera(key: string): boolean {
-
-        if (key == 'd') {
-            this.cameraPosition[0] = this.cameraPosition[0] + 0.05;
-            return true;
-        }
-        else if (key == 'a') {
-            this.cameraPosition[0] = this.cameraPosition[0] - 0.05;
-            return true;
-        }
-        else if (key == 'w') {
-            this.cameraPosition[1] = this.cameraPosition[1] + 0.05;
-            return true;
-        }
-        else if (key == 's') {
-            this.cameraPosition[1] = this.cameraPosition[1] - 0.05;
-            return true;
-        }
-
-        return false;
-    }
-
-    reloadPixelShader(): void {
-
-        if (!this.gl) {
-            return;
-        }
-
-        var gl: WebGLRenderingContext = this.gl;
-
-
-        var vertShader = gl.createShader(gl.VERTEX_SHADER);
-        if (!vertShader) {
-            return;
-        }
-
-        gl.shaderSource(vertShader, this.vertexShaderCode);
-        gl.compileShader(vertShader);
-
-        const codeElement: HTMLElement | null = document.getElementById("code");
-
-        if (!codeElement || !(codeElement instanceof HTMLTextAreaElement)) {
-            return;
-        }
-
-        var fragmentShaderCode = codeElement.value;
-
-        var fragShader = gl.createShader(gl.FRAGMENT_SHADER);
-
-        if (!fragShader) {
-            return;
-        }
-
-        gl.shaderSource(fragShader, fragmentShaderCode);
-        gl.compileShader(fragShader);
-
-        if (!fragShader) {
-            return;
-        }
-
-        this.shaderProgram = gl.createProgram();
-
-        if (!this.shaderProgram) {
-            return;
-        }
-
-        gl.attachShader(this.shaderProgram, vertShader);
-        gl.attachShader(this.shaderProgram, fragShader);
-        gl.linkProgram(this.shaderProgram);
-        gl.useProgram(this.shaderProgram);
-        this.lightingPageBindShaders();
-        this.renderCycle();
-    }
-
-    getCodeSnippet(): string {
-
-        const codeSection: HTMLElement | null = document.getElementById("code");
-
-        if (codeSection && codeSection instanceof HTMLTextAreaElement) {
-            return codeSection.value;
-        }
-
-        return "";
-    }
-
-    reloadVertexShader() {
-
-        if (!this.gl) {
-            return;
-        }
-
-        var gl: WebGLRenderingContext = this.gl;
-
-        var vertexShaderCode = this.getCodeSnippet();
-        var vertShader = gl.createShader(gl.VERTEX_SHADER);
-
-
-        if (!vertShader) {
-            return;
-        }
-
-
-        gl.shaderSource(vertShader, vertexShaderCode);
-        gl.compileShader(vertShader);
-
-        var fragShader = gl.createShader(gl.FRAGMENT_SHADER);
-
-        if (!fragShader) {
-            return;
-        }
-
-
-        gl.shaderSource(fragShader, this.fragmentShaderCode);
-        gl.compileShader(fragShader);
-
-
-        this.shaderProgram = gl.createProgram();
-
-        if (!this.shaderProgram) {
-            return;
-        }
-
-        gl.attachShader(this.shaderProgram, vertShader);
-        gl.attachShader(this.shaderProgram, fragShader);
-        gl.linkProgram(this.shaderProgram);
-        gl.useProgram(this.shaderProgram);
-
-        var uCamPosLocation = gl.getUniformLocation(this.shaderProgram, "u_cameraPosition");
-        gl.uniform3f(uCamPosLocation,
-            this.cameraPosition[0],
-            this.cameraPosition[1],
-            this.cameraPosition[2]
-        );
-
-        this.renderCycle();
-    }
-
-    cameraPageBindShaders() {
-
-        this.bind_a_position();
-        this.bind_a_normal();
-        this.renderCycle();
-    }
-
-    lightingPageBindShaders() {
-
-        this.bind_a_position();
-        this.bind_a_normal();
-        this.renderCycle();
-    }
-
-    loadingPageBindShaders() {
-
-        this.bind_a_position();
-        this.bind_a_colour();
-        this.renderCycle();
-    }
-
-    reloadBuffers() {
-
-        var textAreaContent = this.getCodeSnippet();
-        this.vertices = this.toNumArray(textAreaContent.split(","));
-        this.loadBuffers();
-        this.loadingPageBindShaders();
-    }
-
-    setupCameraMovement() {
-
-        document.addEventListener('keyup', (event) => {
-            this.updateCameraPositionOnKeyUp(event);
-        }, false);
-    }
-
-    public startRotationLoop() {
-
-        window.requestAnimationFrame(() => { this.updateRotation() });
-    }
-
-    public updateRotation() {
-
-        if (this.gl) {
-
-            var gl: WebGLRenderingContext = this.gl;
-
-            if (this.shaderProgram) {
-
-                var uRotationLocation = gl.getUniformLocation(this.shaderProgram, "u_rotation");
-                gl.uniform1f(uRotationLocation, this.rotation);
-                this.rotation = this.rotation + 0.01;
-                if (this.rotation > 6.28) {
-                    this.rotation = this.rotation - 6.28;
-                }
-            }
-        }
-
-        this.renderCycle();
-
-        window.requestAnimationFrame(() => { this.updateRotation() });
-    }
-}
-
-
-class PageBuilder {
-
-    private static getHeader(page: string): string {
-        var header = '\
-        <div id="wrapper">\
-            <h1 style=\"padding:15px;">Code Along at https://tinyurl.com/ct5nrhtd</h1>\
-            <div id="nav">\
-            <ul>';
-        var links = new Array("Setting the Scene", "index.html", "Graphics Pipeline", "pipeline.html", "Loading Data", "loading.html", "Camera Control", "camera.html", "Lighting", "lighting.html");
-        for (var i = links.length - 2; i >= 0; i -= 2) {
-            header = header + '<li';
-            if (links[i + 1] == page) {
-                header = header + ' class="selected"';
-            }
-            header = header + '>\
-                <a href="' + links[i + 1] + '">' + links[i] + '</a>\
-                </li>';
-        }
-        header = header + '\
-            </ul>\
-            </div>\
-            <div id="content">';
-        return header;
-    }
-
-
-    public static writeHeaderAndCanvas(page: string): void {
-
-        const headerString = this.getHeader(page);
-        document.write(headerString);
-
-        const canvasString = '<canvas width="500" height="500" id="webGLCanvas"></canvas>';
-        document.write(canvasString);
-
-        this.writeForm();
-    }
-
-    public static writePlainHeader(page: string): void {
-
-        document.write(this.getHeader(page));
-    }
-
-    private static writeForm(): void {
-
-        var formText = '\
-        <form action="javascript: donothing()">\
-            <textarea id="code" name="vs" rows="15" cols="50"></textarea>\
-                <input type="submit" value="Update" id="update-button">\
-        </form>';
-
-        document.write(formText);
-    }
-
-    public static writeFooter(): void {
-
-        var footerDiv = '</div><div id="footer"><p><b>Simon Grey</b> - <i>S.Grey@hull.ac.uk</i></p></div></div>';
-        document.write(footerDiv);
-    }
-}
-
-
-
-class PageLoader {
-
-    public static initialiseLightingTutorial(vertices: number[], indices: number[]) {
-
-        var vertexShaderCode: string =
-            'attribute vec3 a_position;\r\n' +
-            'attribute vec3 a_normal;\r\n\r\n' +
-            'uniform float u_rotation;\r\n\r\n' +
-            'varying vec3 v_normal;\r\n\r\n' +
-            'void main(void) {\r\n' +
-            'float c = cos(u_rotation);' +
-            'float s = sin(u_rotation);' +
-            'mat4 rot = mat4 (c,0,-s,0, 0,1,0,0, s,0,c,0, 0,0,0,1);' +
-            ' gl_Position = rot * vec4(a_position, 1.0);\r\n' +
-            ' v_normal = a_normal;\r\n' +
-            '}';
-
-        var fragmentShaderCode: string =
-            'precision mediump float;\r\n\r\n' +
-            'varying vec3 v_normal;\r\n\r\n' +
-            'void main(void) {\r\n' +
-            ' gl_FragColor = vec4(v_normal * 0.5 + 0.5, 1.0);\r\n' +
-            '}';
-
-        const codeSection: HTMLElement | null = document.getElementById("code");
-
-        if (codeSection && codeSection instanceof HTMLTextAreaElement) {
-
-            codeSection.value = fragmentShaderCode;
-            var host = new WebGlHost(vertices, indices, vertexShaderCode, fragmentShaderCode, "lighting");
-            host.lightingPageBindShaders();
-            host.startRotationLoop();
-        }
-    }
-
-    public static initialiseCameraTutorial(vertices: number[], indices: number[]) {
-
-        var vertexShaderCode =
-            'attribute vec3 a_position;\r\n' +
-            'attribute vec3 a_normal;\r\n\r\n' +
-            'uniform vec3 u_cameraPosition;\r\n\r\n' +
-            'varying vec4 v_colour;\r\n\r\n' +
-            'void main(void) {\r\n' +
-            ' gl_Position = vec4(a_position, 1.0);\r\n' +
-            ' v_colour = vec4(a_normal * 0.5 + 0.5, 1.0);\r\n' +
-            '}';
-
-        var fragmentShaderCode =
-            'precision mediump float;' +
-            'varying vec4 v_colour;' +
-            'void main(void) {' +
-            ' gl_FragColor = v_colour;' +
-            '}';
-
-        const codeSection: HTMLElement | null = document.getElementById("code");
-
-        if (codeSection && codeSection instanceof HTMLTextAreaElement) {
-
-            codeSection.value = vertexShaderCode;
-            var host = new WebGlHost(vertices, indices, vertexShaderCode, fragmentShaderCode, "camera");
-            host.cameraPageBindShaders();
-            host.setupCameraMovement();
-        }
-    }
-
-    public static initialiseVBOTutorial(vertices: number[], indices: number[]) {
-
-        var vertexShaderCode =
-            'attribute vec3 a_position;' +
-            'attribute vec3 a_colour;' +
-            'varying vec4 v_colour;' +
-            'void main(void) {' +
-            ' gl_Position = vec4(a_position, 1.0);' +
-            ' v_colour = vec4(a_colour, 1.0);' +
-            '}';
-
-        var fragmentShaderCode =
-            'precision mediump float;' +
-            'varying vec4 v_colour;' +
-            'void main(void) {' +
-            ' gl_FragColor = v_colour;' +
-            '}';
-
-        const codeSection: HTMLElement | null = document.getElementById("code");
-        const startingCode: string = "0.0, 0.75, 0.0, 1.0, 0.0, 0.0,-0.75, -0.75, 0.0, 0.0, 1.0, 0.0, 0.75, -0.75, 0.0, 0.0, 0.0, 1.0";
-
-        if (codeSection && codeSection instanceof HTMLTextAreaElement) {
-
-            codeSection.value = startingCode;
-            var host = new WebGlHost(vertices, indices, vertexShaderCode, fragmentShaderCode, "loading");
-            host.loadingPageBindShaders();
-        }
-    }
-}
-
-
-
-function LoadBunnyPage(): void {
-
-    const bunnyModel: StanfordBunny = new StanfordBunny();
-    PageLoader.initialiseLightingTutorial(bunnyModel.bunnyVertices, bunnyModel.bunnyIndices);
-}
-
-
-function loadCameraPage(): void  {
-
-    const sphere: SphereModel = new SphereModel();
-    PageLoader.initialiseCameraTutorial(sphere.sphereVertices, sphere.sphereIndices);
-}
-
-
-function LoadLoadingPage(): void  {
-
-    const triangle: TriangleModel = new TriangleModel();
-    PageLoader.initialiseVBOTutorial(triangle.vertices, triangle.indices);
-}
-
