+++ src/models/triangle.ts
@@ -1,0 +1,6 @@
+
+export class TriangleModel {
+
+    public indices: number[] = [0, 1, 2];
+    public vertices: number[] = [0.0, 0.75, 0.0, 1.0, 0.0, 0.0, -0.75, -0.75, 0.0, 0.0, 1.0, 0.0, 0.75, -0.75, 0.0, 0.0, 0.0, 1.0];
+}
