+++ src/script_loading.ts
@@ -1,0 +1,11 @@
+
+import { TriangleModel } from "./models/triangle.js";
+import { PageLoader } from "./pageloader.js";
+
+function LoadLoadingPage(): void  {
+
+    const triangle: TriangleModel = new TriangleModel();
+    PageLoader.initialiseVBOTutorial(triangle.vertices, triangle.indices);
+}
+
+LoadLoadingPage();
